+++ src/core_modules/capture-core/components/D2Form/D2Section.container.ts
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import { D2Section } from './D2Section.component';
-import type { Section as MetaDataSection } from '../../metaData';
+import { Section as MetaDataSection } from '../../metaData';
 import { withApiUtils } from '../../HOC';
 
 const mapStateToProps = (state: any, props: { sectionMetaData: MetaDataSection, formId: string }) => {
@@ -12,7 +12,10 @@ const mapStateToProps = (state: any, props: { sectionMetaData: MetaDataSection, 
         return { isHidden: !visibleFields.length };
     }
 
-    return { isHidden: !props.sectionMetaData.elements.size };
+    return {
+        isHidden: !props.sectionMetaData.elements.size ||
+            props.sectionMetaData.id === MetaDataSection.LEFTOVERS_SECTION_ID,
+    };
 };
 
 const mapDispatchToProps = () => ({});