+++ src/core_modules/capture-core/components/D2Form/field/Components/OptionSetField/OptionSetSelectFieldForForm.component.ts
@@ -10,6 +10,7 @@ import {
     withDisplayMessages,
     withFilterProps,
     withOptionsIconElement,
+    withSearchHelpMessage,
 } from '../../../../FormFields/New';
 import {
     withRequiredFieldCalculation,
@@ -36,10 +37,12 @@ export const OptionSetSelectFieldForForm = withGotoInterface()(
                                     `${props.fieldLabelMediaBasedClass} ${labelTypeClasses.textLabel}`,
                             })(
                                 withDisplayMessages()(
-                                    withOptionsIconElement()(
-                                        withRulesOptionVisibilityHandler()(
-                                            withFilterProps(getFilteredProps)(
-                                                withSelectSingleTranslations()(SingleSelectField),
+                                    withSearchHelpMessage()(
+                                        withOptionsIconElement()(
+                                            withRulesOptionVisibilityHandler()(
+                                                withFilterProps(getFilteredProps)(
+                                                    withSelectSingleTranslations()(SingleSelectField),
+                                                ),
                                             ),
                                         ),
                                     ),