+++ src/core_modules/capture-core/components/DataEntry/withSaveHandler/withSaveHandler.tsx
@@ -280,6 +280,7 @@ const getSaveHandler = (
                 calculatedFoundation: foundation,
                 sectionsInitialised: reduxSections && Object.values(reduxSections)
                     .every(({ valid }: any) => valid !== undefined),
+                completeDataEntryFieldValue: state.dataEntriesFieldsValue[key]?.complete,
             };
         };