+++ src/core_modules/capture-core/components/DataEntryWidgetOutput/DataEntryWidgetOutput.component.tsx
@@ -6,6 +6,7 @@ import { ErrorsSection } from './ErrorsSection/ErrorsSection.container';
 import { WarningsSection } from './WarningsSection/WarningsSection.container';
 import { WidgetFeedback } from '../WidgetFeedback';
 import { WidgetIndicator } from '../WidgetIndicator';
+import { useHideWidgetByRuleLocations } from '../../hooks';
 
 type OwnProps = {
     onLink: (teiId: string, values: Record<string, unknown>) => void;
@@ -23,35 +24,57 @@ const getStyles = (): Readonly<any> => ({
     },
 });
 
-type Props = OwnProps & WithStyles<typeof getStyles>;
+type Props = OwnProps & WithStyles<typeof getStyles> & {
+    programRules?: Array<any>;
+};
 
-const componentContainers = [
+type ComponentContainer = {
+    id: string;
+    Component: React.ComponentType<any>;
+    shouldHideWidget?: (props: Record<string, any>) => boolean;
+};
+
+const componentContainers: Array<ComponentContainer> = [
     { id: 'ErrorsSection', Component: ErrorsSection },
     { id: 'WarningsSection', Component: WarningsSection },
-    { id: 'FeedbacksSection', Component: WidgetFeedback },
-    { id: 'IndicatorsSection', Component: WidgetIndicator },
+    {
+        id: 'WidgetFeedback',
+        Component: WidgetFeedback,
+        shouldHideWidget: ({ hideWidgets }: any) => hideWidgets?.feedback,
+    },
+    {
+        id: 'WidgetIndicator',
+        Component: WidgetIndicator,
+        shouldHideWidget: ({ hideWidgets }: any) => hideWidgets?.indicator,
+    },
 ];
 
-class DataEntryWidgetOutputPlain extends React.Component<Props> {
-    renderComponent = (
-        container: { id: string, Component: React.ComponentType<any> },
-        props: Record<string, any>,
-    ) => <container.Component key={container.id} {...props} />
+const DataEntryWidgetOutputPlain = (props: Props) => {
+    const { classes, programRules, ...passOnProps } = props;
 
-    render() {
-        const { classes, ...passOnProps } = this.props;
-        return (
-            <StickyOnScroll
-                offsetTop={50}
-                minViewpointWidth={768}
-                containerClass={classes.stickyOnScroll}
-            >
-                <div className={classes.container}>
-                    {componentContainers.map(c => this.renderComponent(c, passOnProps))}
-                </div>
-            </StickyOnScroll>
-        );
-    }
-}
+    const hideWidgets = useHideWidgetByRuleLocations(programRules || []);
+
+    const renderComponent = (
+        container: ComponentContainer,
+        componentProps: Record<string, any>,
+    ) => {
+        const { shouldHideWidget } = container;
+        const hideWidget = shouldHideWidget?.(componentProps);
+        if (hideWidget) return null;
+        return <container.Component key={container.id} {...componentProps} />;
+    };
+
+    return (
+        <StickyOnScroll
+            offsetTop={50}
+            minViewpointWidth={768}
+            containerClass={classes.stickyOnScroll}
+        >
+            <div className={classes.container}>
+                {componentContainers.map(c => renderComponent(c, { ...passOnProps, hideWidgets }))}
+            </div>
+        </StickyOnScroll>
+    );
+};
 
 export const DataEntryWidgetOutputComponent = withStyles(getStyles)(DataEntryWidgetOutputPlain);