+++ src/core_modules/capture-core/components/FormFields/New/index.ts
@@ -23,6 +23,7 @@ export { DateTimeRangeField } from './Fields/DateAndTimeFields/DateTimeRangeFiel
 export { withCalculateMessages } from './HOC/messages/withCalculateMessages';
 export { withConvertedOptionSet } from './HOC/withConvertedOptionSet';
 export { withDisplayMessages } from './HOC/messages/withDisplayMessages';
+export { withSearchHelpMessage } from './HOC/messages/withSearchHelpMessage';
 export { withDefaultFieldContainer } from './HOC/withDefaultFieldContainer';
 export { withFilterProps } from './HOC/withFilterProps';
 export { withGotoInterface } from './HOC/withGotoInterface';
@@ -32,6 +33,7 @@ export { withStyledContainer } from './HOC/withStyledContainer';
 export { withOptionsIconElement } from './HOC/withOptionsIconElement';
 export { withFocusSaver, withInternalChangeHandler } from 'capture-ui';
 export { withCenterPoint } from './HOC/withCenterPoint';
+export { withConditionalTooltip } from './HOC/withConditionalTooltip';
 
 // OrgUnit HOCs
 export {