+++ src/core_modules/capture-core/components/Pages/ViewEvent/RightColumn/FeedbacksSection/FeedbacksSection.types.ts
@@ -1,6 +1,0 @@
-export type PlainProps = {
-    feedbacks?: {
-        displayTexts?: Array<string> | null;
-        displayKeyValuePairs?: Array<{ key: string; value: string }> | null;
-    } | null;
-};