+++ src/core_modules/capture-core/metaDataMemoryStoreBuilders/programs/factory/enrollment/EnrollmentFactory.ts
@@ -136,6 +136,33 @@ export class EnrollmentFactory {
         return section;
     }
 
+    async _addLeftoversSection(
+        enrollmentForm: RenderFoundation,
+        cachedProgramTrackedEntityAttributes: Array<CachedProgramTrackedEntityAttribute> | null,
+    ) {
+        if (!cachedProgramTrackedEntityAttributes) return;
+
+        // Check if there exist attributes which are not assigned to a section
+        const attributesInSection = enrollmentForm.getElements().reduce((acc, attribute) => {
+            acc.add(attribute.id);
+            return acc;
+        }, new Set());
+
+        const unassignedAttributes = cachedProgramTrackedEntityAttributes
+            .filter(attribute => !attributesInSection.has(attribute.trackedEntityAttributeId));
+
+        if (unassignedAttributes.length === 0) return;
+
+        // Create a special section for the unassigned attributes
+        const section = new Section((o) => {
+            o.id = Section.LEFTOVERS_SECTION_ID;
+            o.group = Section.groups.ENROLLMENT;
+        });
+
+        await this._buildElementsForSection(unassignedAttributes, section);
+        enrollmentForm.addSection(section);
+    }
+
     async _buildElementsForSection(
         cachedProgramTrackedEntityAttributes: Array<CachedProgramTrackedEntityAttribute>,
         section: Section,
@@ -332,6 +359,9 @@ export class EnrollmentFactory {
             );
             section && enrollmentForm.addSection(section);
         }
+
+        await this._addLeftoversSection(enrollmentForm, cachedProgramTrackedEntityAttributes);
+
         return enrollmentForm;
     }
 
@@ -349,6 +379,7 @@ export class EnrollmentFactory {
             o.disabled = searchGroupElement.disabled;
             o.type = teiAttribute.valueType;
             o.optionSet = searchGroupElement.optionSet;
+            o.searchOperator = searchGroupElement.searchOperator;
         });
         return element;
     }